Add Medium Java Solution for 4248. Count Commas in Range II from LeetCode

+++ LeetCode/Patterns/Math/4248. Count Commas in Range II/Solution.java
@@ -16,6 +16,6 @@ class Solution {
             start *= 1000;
             commas++;
         }
-        return (int) ans ;
+        return ans ;
     }
 }
